Initial Commit - This project is an exploration of pandas package functionality through a movie recommendation system based on content.
[Long] - This software using sklearn and pandas extensions. It takes the data from an online database and transforms it into a dataframe, it transforms the dataframe into a matrix and calculates the cosine similarity.

The final result is one where you can input a specific movie id (in the current case id=7) and the output in the command line will be the top recommended movies based on their similarity in content.

There are updates to be made. The README.md has to be expanded and description of what I have found in this project need to be made.